Add support for the TTGO T-Beam V1
This adds support for the TTGO T-Beam LoRa board by building upon
the LoPy4 implementation.

For supporting this board, the safeboot pin has been changed to GPIO25,
because on the T-Beam V1, IO21 is pulled up to 3v3 causing a safe boot.

Also, disable the RGB LED for this board.

https://github.com/nunomcruz/pycom-micropython-sigfox/tree/tbeamv1-support